demo/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
llama-index = "^0.10.26"
load-dotenv = "^0.1.0"
qdrant-client = "^1.8.2"
llama-index-vector-stores-qdrant = "^0.1.6"
llama-index-postprocessor-voyageai-rerank = "^0.1.1"
langchain-openai = "^0.1.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
